move boundary conditions inside linkedcell
