+++ src/io/input/cli/SimParams.h
@@ -1,4 +1,4 @@
-#include "simulator/particle/BoundaryCondition.h"
+#include "simulator/particle/container/LinkedCellContainer.h"
 #include <array>
 #include <string>
 #pragma once
